24332: Modify LatitudeLongitudeIssuesController update

+++ lib/roo_helper/loader.rb
@@ -43,7 +43,7 @@ module RooHelper
         sheet_klass = sheet_options[:klass]
 
         sheet_klass.transaction do
-          sheet_klass.delete_all
+          sheet_klass.delete_all unless file_options[:skip_loading]
 
           processed_sheet = if %w[.xls .xlsx].include?(ext) && parse_as_xml?(sheet, index)
                               process_as_xml(sheet_klass, sheet, index, sheet_options)
@@ -51,13 +51,8 @@ module RooHelper
                               process_sheet(sheet_klass, sheet, sheet_options, file_options)
                             end
 
-          results = if file_options[:skip_loading]
-                      processed_sheet[:results]
-                    else
-                      sheet_klass.load_records(processed_sheet[:results], sheet_options)
-          end
           loaded_sheets << {
-            results: results,
+            results: sheet_klass.load_records(processed_sheet[:results], sheet_options),
             header_warnings: processed_sheet[:header_warnings],
             klass: sheet_klass
           }
